# doc/api.yml
openapi: 3.0.3
info:
  title: Library Management API
  description: >
    REST API for managing books, authors, genres, and book copies.
    Supports full CRUD, paginated listings, and multi-criteria book search.
  version: 1.0.0
servers:
  - url: https://api.library.com/v1
    description: Production server
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List all customers
      operationId: listCustomers
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: Paginated list of customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPage'

    post:
      tags: [Customers]
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreation'
      responses:
        '201':
          description: Customer successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
  /customers/{id}:
    parameters:
      - $ref: '#/components/parameters/id'

    get:
      tags: [Customers]
      summary: Get a customer by ID
      operationId: getCustomer
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Customers]
      summary: Update a customer
      operationId: updateCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: Customer successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Customers]
      summary: Delete a customer
      operationId: deleteCustomer
      responses:
        '204':
          description: Customer successfully deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Sales ─────────────────────────────────────
  /sales:
    get:
      tags: [Sales]
      summary: List all sales
      operationId: listSales
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: Paginated list of sales
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalePage'

    post:
      tags: [Sales]
      summary: Create a new sale
      operationId: createSale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleCreation'
      responses:
        '201':
          description: Sale successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          $ref: '#/components/responses/BadRequest'
  /sales/{id}:
    parameters:
      - $ref: '#/components/parameters/id'

    get:
      tags: [Sales]
      summary: Get sale details by ID
      operationId: getSale
      responses:
        '200':
          description: Sale found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Sales]
      summary: Update a sale
      operationId: updateSale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleUpdate'
      responses:
        '200':
          description: Sale successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Sales]
      summary: Delete a sale
      operationId: deleteSale
      responses:
        '204':
          description: Sale successfully deleted
        '404':
          $ref: '#/components/responses/NotFound'

# ─────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────
components:
...
  parameters:
    id:
      name: id
      in: path
      required: true
      description: Unique resource identifier (UUID)
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"
    page:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    pageSize:
      name: pageSize
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # ── Reusable responses ────────────────────────
  responses:
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404
            message: "Resource not found"
    BadRequest:
      description: The request body or parameters are invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: "Validation error: 'title' is required"

  # ── Schemas ───────────────────────────────────
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Human-readable error description
    PaginationMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Number of items per page
          example: 20
        totalItems:
          type: integer
          description: Total number of items matching the query
          example: 142
        totalPages:
          type: integer
          description: Total number of pages
          example: 8

    # NOTE: HARD_BDACK in the original source is a typo — corrected to HARD_BACK
...
    Customer:
      type: object
      required: [id, fullName, email]
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-4a5b-8c9d-0e1f2a3b4c5d"
        fullName:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
    CustomerCreation:
      type: object
      required: [fullName, email]
      properties:
        fullName:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
    CustomerUpdate:
      type: object
      required: [fullName, email]
      properties:
        fullName:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
    CustomerPage:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # ── Sale ─────────────────────────────────
    Sale:
      type: object
      required: [id, customer, saleDate, status, paymentMethod, totalAmount, items]
      properties:
        id:
          type: string
          format: uuid
          example: "e1f2a3b4-c5d6-4a5b-8c9d-0e1f2a3b4c5d"
        customer:
          $ref: '#/components/schemas/Customer'
        saleDate:
          type: string
          format: date
          example: "2024-06-21"
        status:
          $ref: '#/components/schemas/SaleStatus'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        totalAmount:
          type: number
          format: float
          minimum: 0
          example: 29.98
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleItem'
    SaleItem:
      type: object
      required: [id, bookCopyId, unitPrice, quantity]
      properties:
        id:
          type: string
          format: uuid
        bookCopyId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        unitPrice:
          type: number
          format: float
          minimum: 0
          example: 14.99
        quantity:
          type: integer
          minimum: 1
          example: 2
    SaleItemRequest:
      type: object
      required: [bookCopyId, quantity]
      properties:
        bookCopyId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        quantity:
          type: integer
          minimum: 1
          example: 2
    SaleCreation:
      type: object
      required: [customerId, saleDate, paymentMethod, items]
      properties:
        customerId:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-4a5b-8c9d-0e1f2a3b4c5d"
        saleDate:
          type: string
          format: date
          example: "2024-06-21"
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        status:
          $ref: '#/components/schemas/SaleStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleItemRequest'
    SaleUpdate:
      type: object
      description: All fields are optional — only provided fields will be updated
      properties:
        saleDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/SaleStatus'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
    SalePage:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Sale'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    SaleStatus:
      type: string
      enum:
        - PAID
        - RESERVED
        - SHIPPED
        - CANCELLED
        - PENDING
    PaymentMethod:
      type: string
      enum:
        - CARD
        - CASH
        - MOBILE_MONEY